feat: add checkPermissions in all places checkPermissionsLegacy is used

+++ src/plugins/shipping-rates/mutations/deleteFlatRateFulfillmentRestriction.js
@@ -17,10 +17,11 @@ export default async function deleteFlatRateFulfillmentRestriction(context, inpu
   inputSchema.validate(input);
 
   const { restrictionId, shopId } = input;
-  const { checkPermissionsLegacy, collections } = context;
+  const { checkPermissions, checkPermissionsLegacy, collections } = context;
   const { FlatRateFulfillmentRestrictions } = collections;
 
   await checkPermissionsLegacy(["admin", "owner", "shipping"], shopId);
+  await checkPermissions(`reaction:shippingRestriction:${restrictionId}`, "delete", { shopId });
 
   const { ok, value } = await FlatRateFulfillmentRestrictions.findOneAndDelete({
     _id: restrictionId,
