+++ src/plugins/surcharges/mutations/deleteSurcharge.js
@@ -17,10 +17,11 @@ export default async function deleteSurcharge(context, input) {
   inputSchema.validate(input);
 
   const { surchargeId, shopId } = input;
-  const { checkPermissionsLegacy, collections } = context;
+  const { checkPermissions, checkPermissionsLegacy, collections } = context;
   const { Surcharges } = collections;
 
   await checkPermissionsLegacy(["admin", "owner", "shipping"], shopId);
+  await checkPermissions(`reaction:surcharge:${surchargeId}`, "delete", { shopId });
 
   const { ok, value } = await Surcharges.findOneAndDelete({
     _id: surchargeId,
